Fixes niche representation
I realised that the niche number was not reflective of the actual niche number, since the null values were not included, so I had to change the code to include also null values in the legend so that the numbering would still be accurate. I hereby also simplified the code a bit, since getting rid of the null values was no longer as necessary.